yamlet-backend/requirements.txt
pyspark>=3.5.5
pyyaml>=6.0.2
mysql-connector-python>=9.2.0
pytest>=8.3.5
fastapi>=0.115.12
uvicorn>=0.34.0
mypy>=1.15.0
python-crontab>=3.2.0
python-multipart>=0.0.20
crontab